feat: split functions relate to Tour
