+++ src/Service/TourPlanService.php
@@ -1,0 +1,66 @@
+<?php
+
+namespace App\Service;
+
+use App\Entity\Tour;
+use App\Entity\TourPlan;
+use App\Repository\DestinationRepository;
+use App\Repository\TourPlanRepository;
+use App\Request\TourRequest;
+use App\Request\TourUpdateRequest;
+
+class TourPlanService
+{
+    private DestinationRepository $destinationRepository;
+    private TourPlanRepository $tourPlanRepository;
+
+    public function __construct(
+        DestinationRepository   $destinationRepository,
+        TourPlanRepository      $tourPlanRepository,
+    )
+    {
+        $this->destinationRepository = $destinationRepository;
+        $this->tourPlanRepository = $tourPlanRepository;
+    }
+
+    public function addTourPlan(TourRequest $tourRequest, Tour $tour): Tour
+    {
+        foreach ($tourRequest->getTourPlans() as $tourPlanRequest) {
+            $destination = $this->destinationRepository->find($tourPlanRequest['destination']);
+            if (!is_object($destination)) {
+                continue;
+            }
+            $tourPlan = new TourPlan();
+            $tourPlan->setTitle($tourPlanRequest['title']);
+            $tourPlan->setDescription($tourPlanRequest['description']);
+            $tourPlan->setDay($tourPlanRequest['day']);
+            $tourPlan->setDestination($destination);
+            $tourPlan->setTour($tour);
+            $this->tourPlanRepository->add($tourPlan);
+        }
+
+        return $tour;
+    }
+
+    public function updateTourPlan(TourUpdateRequest $tourUpdateRequest): void
+    {
+        foreach ($tourUpdateRequest->getTourPlans() as $tourPlanRequest) {
+            $destination = $this->destinationRepository->find($tourPlanRequest['destination']);
+            if (!is_object($destination)) {
+
+                continue;
+            }
+            $tourPlan = $this->tourPlanRepository->find($tourPlanRequest['id']);
+            if (!is_object($tourPlan)) {
+
+                continue;
+            }
+            $tourPlan->setTitle($tourPlanRequest['title'] ?? $tourPlan->getTitle());
+            $tourPlan->setDescription($tourPlanRequest['description'] ?? $tourPlan->getDescription());
+            $tourPlan->setDay($tourPlanRequest['day'] ?? $tourPlan->getDay());
+            $tourPlan->setDestination($destination ?? $tourPlan->getDestination());
+            $tourPlan->setUpdatedAt(new \DateTimeImmutable());
+            $this->tourPlanRepository->add($tourPlan);
+        }
+    }
+}
